Move ClassHandle and related types out of gui package
