+++ enigma-swing/src/main/java/cuchaz/enigma/gui/GuiController.java
@@ -37,8 +37,9 @@ import cuchaz.enigma.gui.config.Config;
 import cuchaz.enigma.gui.dialog.ProgressDialog;
 import cuchaz.enigma.gui.stats.StatsGenerator;
 import cuchaz.enigma.gui.stats.StatsMember;
-import cuchaz.enigma.gui.util.ClassHandle;
+import cuchaz.enigma.classhandle.ClassHandle;
 import cuchaz.enigma.gui.util.History;
+import cuchaz.enigma.classhandle.ClassHandleProvider;
 import cuchaz.enigma.network.*;
 import cuchaz.enigma.network.packet.LoginC2SPacket;
 import cuchaz.enigma.network.packet.Packet;
