+++ enigma-swing/src/main/java/cuchaz/enigma/gui/panels/PanelEditor.java
@@ -26,12 +26,12 @@ import cuchaz.enigma.gui.Gui;
 import cuchaz.enigma.gui.GuiController;
 import cuchaz.enigma.gui.config.Themes;
 import cuchaz.enigma.gui.elements.PopupMenuBar;
-import cuchaz.enigma.gui.events.ClassHandleListener;
+import cuchaz.enigma.events.ClassHandleListener;
 import cuchaz.enigma.gui.events.EditorActionListener;
 import cuchaz.enigma.gui.events.ThemeChangeListener;
 import cuchaz.enigma.gui.highlight.BoxHighlightPainter;
 import cuchaz.enigma.gui.highlight.TokenHighlightType;
-import cuchaz.enigma.gui.util.ClassHandle;
+import cuchaz.enigma.classhandle.ClassHandle;
 import cuchaz.enigma.gui.util.ScaleUtil;
 import cuchaz.enigma.source.Token;
 import cuchaz.enigma.translation.mapping.EntryRemapper;
